Create metal_d_orbital_finder.py
This python script is not run directly but with help of a driver script written in bash (look for driver.sh in My-codes). The purpose of the script is to find the molecular orbitals which have significant d-orbital character. The input comes from Loewdin orbital decomposition analysis generated by an Orca calculation.  This script is still in development phase, and I will soon upload a balck-box code which will enable you to generate good starting orbitals for CASSCF calculations using ORCA.